refactor: move data methods to parent validation component
